requirements.txt
scipy
scikit-learn
numpy
matplotlib
seaborn
opencv_python
Pillow
torch>=1.2.0
torchvision
